feat: 创建完整的TypeScript版使用文档
- 创建完整的使用文档 (README.md)
- 创建API参考文档 (API.md)
- 创建示例代码文档 (EXAMPLES.md)
- 创建性能优化文档 (PERFORMANCE.md)
- 创建故障排除文档 (TROUBLESHOOTING.md)
- 创建迁移指南文档 (MIGRATION.md)
- 创建贡献指南文档 (CONTRIBUTING.md)

文档包含:
- 快速开始指南
- 完整的API参考
- 丰富的示例代码
- 性能优化技巧
- 故障排除指南
- 从其他库迁移指南
- 贡献指南

所有文档都提供了详细的代码示例和最佳实践。

+++ src/filter_pack.rs
@@ -22,7 +22,7 @@ impl FilterPacker {
         let bytes_per_row = (width as usize * bpp) as usize;
         
         for y in 0..height {
-            let row_start = (y * bytes_per_row) as usize;
+            let row_start = (y * bytes_per_row as u32) as usize;
             let row_end = row_start + bytes_per_row;
             
             if row_end > data.len() {
@@ -152,10 +152,10 @@ impl FilterUnpacker {
                 row_index: y as usize,
                 column_index: 0,
                 previous_row: if y > 0 {
-                    let prev_start = (y - 1) * bytes_per_row;
-                    let prev_end = prev_start + bytes_per_row;
-                    if prev_end <= unpacked_data.len() {
-                        Some(unpacked_data[prev_start..prev_end].to_vec())
+                    let prev_start = (y - 1) * bytes_per_row as u32;
+                    let prev_end = prev_start + bytes_per_row as u32;
+                    if prev_end <= unpacked_data.len() as u32 {
+                        Some(unpacked_data[prev_start as usize..prev_end as usize].to_vec())
                     } else {
                         None
                     }
